Add RaceInv class and update relevant dependencies
A new class named RaceInv was added to manage inventory actions and interactions. The "ToastWonderland" library's dependency in the build.gradle.kts file was updated. Changes were made in the Commands and EmpowermentStoneHandler classes to integrate the new class. Also, an 'onCommand' method was added to the ToastRpg class to properly handle command executions.